protocols: refactor/rename software protocols

+++ driver/protocols/protocol.h
@@ -13,11 +13,11 @@ enum protocol_types {
 };
 
 struct protocol_interface {
-	unsigned char (*read_cmd_data)(const unsigned char *cmd, unsigned int cmd_length, unsigned char *data, unsigned int data_length);
-	unsigned char (*read_data)(unsigned char *data, unsigned int length);
+	unsigned char (*read_cmd_data)(const unsigned char *cmd, unsigned short cmd_length, unsigned char *data, unsigned short data_length);
+	unsigned char (*read_data)(unsigned char *data, unsigned short length);
 	unsigned char (*read_byte)(unsigned char *bdata);
-	unsigned char (*write_cmd_data)(const unsigned char *cmd, unsigned int cmd_length, const unsigned char *data, unsigned int data_length);
-	unsigned char (*write_data)(const unsigned char *data, unsigned int length);
+	unsigned char (*write_cmd_data)(const unsigned char *cmd, unsigned short cmd_length, const unsigned char *data, unsigned short data_length);
+	unsigned char (*write_data)(const unsigned char *data, unsigned short length);
 	unsigned char (*write_byte)(unsigned char bdata);
 	enum protocol_types protocol_type;
 };
